Move plugin scripts into plugin dir, symlink from scripts/
Scripts (config_utils.sh, export_transcript_paths.sh, filter_transcript.py,
stop_hook_gates.sh) now live in plugins/imbue-code-review/scripts/ as real
files, with scripts/ containing symlinks to them. Previously the direction
was reversed.

Also fix stop_hook_gates.sh to only mention /verify-conversation in the
hint message when conversation review is actually one of the missing gates.
